+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Arrow-Right icon.
 ### Changed
  - Pager element styling that fits alpaca design.
+ - Adjustements in slider module
 
 ## [0.4.1] - 13.11.2017
 ### Fixed