+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
  - Pager element styling that fits alpaca design.
+ - Swatcher update
 
 ## [0.4.1] - 13.11.2017
 ### Fixed